Add files via upload
Renewable energy is essential to the global energy transition towards a more sustainable and
carbon-neutral future. In Germany, wind energy is a significant renewable energy source,
accounting for a substantial proportion of the country's electricity generation. The relationship
between wind energy generation and electricity prices is of great importance to policymakers,
energy market regulators, and investors 
.
The aim of this study is to examine the correlation between wind energy generation and
electricity prices in Germany. The analysis uses data obtained from the German transmission
system operator, smard.de 
, for the year 2022. Two datasets were obtained: total wind energy
generation and day-ahead electricity prices. The data analysis includes the distribution of
features, correlation analysis, and time series plots . The results provide insights into the
distribution of wind energy generation and electricity prices in Germany and the relationship
between the two. The analysis can help policymakers and investors make informed decisions
about the future of Germany's renewable energy and energy markets.